+++ lib/ui/pages/edit_tempo_form.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:in_the_pocket/bloc/standalone_metronome_bloc.dart';
 import 'package:in_the_pocket/bloc/tempo_bloc.dart';
 import 'package:in_the_pocket/model/setlistdb.dart';
+import 'package:in_the_pocket/ui/controls/metronome.dart';
 import 'package:provider/provider.dart';
 
 class EditTempoForm extends StatefulWidget {
@@ -14,7 +16,15 @@ class EditTempoForm extends StatefulWidget {
 }
 
 class EditTempoFormState extends State<EditTempoForm> {
-  EditTempoFormState(this.tempo);
+  EditTempoFormState(this.tempo) {
+    beatUnitValueMap = beatUnitEntries.asMap().map(
+        (_, DropdownMenuEntry<int> entry) =>
+            MapEntry<int, String>(entry.value, entry.label));
+
+    beatUnitLabelMap = beatUnitEntries.asMap().map(
+        (_, DropdownMenuEntry<int> entry) =>
+            MapEntry<String, int>(entry.label, entry.value));
+  }
 
   Tempo? tempo;
   final TextEditingController _accentBeatsPerBarController =
@@ -24,16 +34,24 @@ class EditTempoFormState extends State<EditTempoForm> {
   final TextEditingController _bpmController = TextEditingController();
   final TextEditingController _numberofBarsController = TextEditingController();
 
-  bool _dottedQuarterAccent = false;
+  final List<DropdownMenuEntry<int>> beatUnitEntries =
+      const <DropdownMenuEntry<int>>[
+    DropdownMenuEntry<int>(value: 2, label: '1/2'),
+    DropdownMenuEntry<int>(value: 4, label: '1/4'),
+    DropdownMenuEntry<int>(value: 8, label: '1/8'),
+    DropdownMenuEntry<int>(value: 16, label: '1/16'),
+  ];
+
+  late Map<int, String> beatUnitValueMap;
+  late Map<String, int> beatUnitLabelMap;
 
   @override
   void initState() {
     _accentBeatsPerBarController.text =
         tempo?.accentBeatsPerBar.toString() ?? '1';
     _beatsPerBarController.text = tempo?.beatsPerBar.toString() ?? '4';
-    _beatUnitController.text = tempo?.beatUnit.toString() ?? '4';
+    _beatUnitController.text = beatUnitValueMap[tempo?.beatUnit ?? 4] ?? '1/4';
     _bpmController.text = tempo?.bpm.toString() ?? '';
-    _dottedQuarterAccent = tempo?.dottedQuarterAccent ?? false;
 
     if (tempo?.numberOfBars == 0 || tempo?.numberOfBars == null) {
       _numberofBarsController.text = '';
@@ -47,6 +65,12 @@ class EditTempoFormState extends State<EditTempoForm> {
   Widget build(BuildContext context) {
     final TempoBloc tempoBloc = Provider.of<TempoBloc>(context);
     final Track track = tempoBloc.track;
+    final StandaloneMetronomeBloc metronomeBloc =
+        Provider.of<StandaloneMetronomeBloc>(context);
+    metronomeBloc.accentBeatsPerBar = tempo?.accentBeatsPerBar ?? 1;
+    metronomeBloc.beatsPerBar = tempo?.beatsPerBar ?? 4;
+    metronomeBloc.beatUnit = tempo?.beatUnit ?? 4;
+    metronomeBloc.bpm = (tempo?.bpm ?? 60).round();
 
     String pageTitle = 'Tempo Info';
     if (track.title != null) {
@@ -60,13 +84,16 @@ class EditTempoFormState extends State<EditTempoForm> {
           onPressed: () {
             final Tempo tempoToSave = tempo ?? Tempo();
 
-            tempoToSave.accentBeatsPerBar =
-                int.tryParse(_accentBeatsPerBarController.text) ?? 60;
-            tempoToSave.beatsPerBar =
-                int.tryParse(_beatsPerBarController.text) ?? 4;
-            tempoToSave.beatUnit = int.tryParse(_beatUnitController.text) ?? 4;
-            tempoToSave.bpm = double.tryParse(_bpmController.text) ?? 60;
-            tempoToSave.dottedQuarterAccent = _dottedQuarterAccent;
+            tempoToSave.accentBeatsPerBar = metronomeBloc.accentBeatsPerBar;
+            tempoToSave.beatsPerBar = metronomeBloc.beatsPerBar;
+            tempoToSave.bpm = metronomeBloc.bpm.toDouble();
+
+            // Hardcode this to false.  We are just using individual pulses for clicks all the time,
+            // so dotted quarter BPM processing for 6/8 isn't relevant.
+            tempoToSave.dottedQuarterAccent = false;
+
+            tempoToSave.beatUnit =
+                beatUnitLabelMap[_beatUnitController.text] ?? 4;
             tempoToSave.numberOfBars =
                 double.tryParse(_numberofBarsController.text) ?? 0;
             tempoToSave.trackId = track.id;
@@ -86,33 +113,14 @@ class EditTempoFormState extends State<EditTempoForm> {
         color: Theme.of(context).scaffoldBackgroundColor,
         child: ListView(
           children: <Widget>[
+            const MetronomeControl(),
             ListTile(
               leading: const Icon(Icons.title),
-              title: TextField(controller: _bpmController),
-              subtitle: const Text('BPM'),
-            ),
-            ListTile(
-              leading: const Icon(Icons.description),
-              title: TextField(controller: _beatsPerBarController),
-              subtitle: const Text('Beats Per Bar'),
-            ),
-            ListTile(
-              leading: const Icon(Icons.title),
-              title: TextField(controller: _beatUnitController),
+              title: DropdownMenu<int>(
+                dropdownMenuEntries: beatUnitEntries,
+                controller: _beatUnitController,
+              ),
               subtitle: const Text('Beat Unit'),
-            ),
-            CheckboxListTile(
-              controlAffinity: ListTileControlAffinity.leading,
-              title: const Text('Dotted Quarter Accent'),
-              value: _dottedQuarterAccent,
-              onChanged: (bool? check) {
-                setState(() => _dottedQuarterAccent = check ?? false);
-              },
-            ),
-            ListTile(
-              leading: const Icon(Icons.title),
-              title: TextField(controller: _accentBeatsPerBarController),
-              subtitle: const Text('Accent Beats Per Bar'),
             ),
             ListTile(
               leading: const Icon(Icons.title),